@mattcary 2025-04-22
This should be easy to add an e2e test for the presubmit?

Let's submit this now though and add e2e later (which I think would be preferable to an internal GKE test).

The e2e test could just be for the simple HdB case, testing thoughput and such gets harder with machine compatibility. 

/lgtm
/approve